\name{latdiag-package}
\alias{latdiag-package}
\alias{draw}
\docType{package}
\title{\packageTitle{latdiag}}
\description{\packageDescription{latdiag}
}
\details{\packageDESCRIPTION{latdiag}
\packageIndices{latdiag}

The package writes a file of commands for subsequent processing by
the \code{dot} program from \code{graphviz}
}
\author{
Michael Dewey

Maintainer: Michael Dewey <lists@dewey.myzen.co.uk>
}
\references{
P R Rosenbaum.
Probability inqualities for latent scales.
\emph{British Journal of Mathematical and Statistical Psychology},
  \bold{40}: 157--168, 1987
}
\keyword{ package }
