\name{latdiag-package}
\alias{latdiag-package}
\alias{draw}
\docType{package}
\title{\packageTitle{latdiag}}
\description{\packageDescription{latdiag}
}
\details{\packageDESCRIPTION{latdiag}
\packageIndices{latdiag}

The package writes a file of commands for subsequent processing by
the \code{dot} program from \code{graphviz}
}
\author{
Michael Dewey

Maintainer: Michael Dewey <lists@dewey.myzen.co.uk>
}
\references{
\insertRef{rosenbaum87}{latdiag}
}
\keyword{ package }
