\name{latdiag-package}
\alias{latdiag-package}
\alias{draw}
\Rdversion{1.1}
\docType{package}
\title{
Draw latent scale diagram
}
\description{
Writes a file of commands for the dot program to draw a graph
proposed by Rosenbaum and useful for checking some properties of various
sorts of latent scale
}
\details{
\tabular{ll}{
Package: \tab latdiag\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2009-02-04\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
The package writes a file of commands for subsequent processing by
the \code{dot} program from \code{graphviz}
}
\author{
Michael Dewey

Maintainer: Michael Dewey <info@aghmed.fsnet.co.uk>
}
\references{
P R Rosenbaum.
Probability inqualities for latent scales.
\emph{British Journal of Mathematical and Statistical Psychology},
  \bold{40}: 157--168, 1987
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
